explicitly cast random number to an unsigned int for client session ID
massive multiplication was executed first using standard signed 32-bit integers. Only after it completely breaks the integer ceiling did it try to apply the & 0xFFFFFF mask. By then, the Undefined Behavior was already triggered.

+++ fpsgame/server.cpp
@@ -2660,7 +2660,7 @@ namespace server
         clientinfo *ci = getinfo(n);
         ci->clientnum = ci->ownernum = n;
         ci->connectmillis = totalmillis;
-        ci->sessionid = (rnd(0x1000000)*((totalmillis%10000)+1))&0xFFFFFF;
+        ci->sessionid = ((unsigned int)rnd(0x1000000) * ((totalmillis % 10000) + 1)) & 0xFFFFFF; 
 
         connects.add(ci);
         if(!m_mp(gamemode)) return DISC_LOCAL;
